requirements.txt
torch
diffusers
transformers
accelerate
Pillow
pyTelegramBotAPI
